% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_forecast.R
\name{constrained_forecast}
\alias{constrained_forecast}
\title{Constrained Forecast of One-sided Integer Response Model}
\usage{
constrained_forecast(Model, lower, upper)
}
\arguments{
\item{Model}{This is the forecast of the scaled fitted values of supported models.}

\item{lower}{The lower limit of the forecast}

\item{upper}{The upper limit of the forecast}
}
\value{
A list of forecast values within 80\% and 95\% confidence band. The values are:
\item{\code{Lower 80\%}}{Forecast at lower 80\% confidence level.}
\item{\code{Upper 80\%}}{Forecast at upper 80\% confidence level.}
\item{\code{Lower 95\%}}{Forecast at lower 95\% confidence level.}
\item{\code{Upper 95\%}}{Forecast at upper 95\% confidence level.}
}
\description{
Constrained Forecast of One-sided Integer Response Model
}
\examples{
X <- 0:35
lower <- 0
upper <- 35
}
