% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplot.R
\name{corplot}
\alias{corplot}
\title{Custom plot of correlation matrix}
\usage{
corplot(r)
}
\arguments{
\item{r}{Correlation matrix of the data for the plot}
}
\value{
The function returns a custom plot of the correlation matrix
\item{\code{corplot}}{The custom plot of the correlation matrix}
}
\description{
This is a custom plot for correlation matrix in which the coefficients are
displayed along with graphics showing the magnitude of each coefficient.
}
