% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_Dyn.R
\name{Dyn4cast-deprecated}
\alias{Dyn4cast-deprecated}
\title{Deprecated functions in Dyn4cast}
\description{
These functions have been removed (defunct) as indicated against their names.
}
\details{
\itemize{
\item \code{\link{Model_factors}}: This function is deprecated as of
11.11.26 and is no loner in use. Use \strong{model_factors} instead.
}
}
