% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicForecast.R
\name{Dyn4cast}
\alias{Dyn4cast}
\title{Dymanic Forecast of five models and thier Essembles}
\arguments{
\item{Data}{A two column (Date, Case) DAILY dataset for the estimation. The date must be in format recognized by R. In future versions provisions shall be made of quarterly, monthly and yearly data}

\item{BREAKS}{A vector of numbers indicating points of breaks for estimation of the spline models}

\item{MaximumDate}{The date indicating the maximum date (last date) in the data frame, meaning that forecasting starts the next date following it. The date must be a recognized date format. Note that for forecasting, the date origin is set to 1970-01-01

utils::globalVariables(c("Date", "Day", "Forecast", "Models"))}
}
\value{
A list with the following components:
\item{\code{Forecast}}{The forecast is equivalent to the length of the dataset (equal days forecast).}
\item{\code{RMSE}}{Root Mean Sqaure Error (rmse) for each forecast.}
\item{\code{Plot}}{The combined plots of the forecasts using ggplot. }
\item{\code{Date}}{This is the date range for the forecast.}
}
\description{
This function estimates, predict and forecast five models and their Essembles. The recognised models are lm, smooth spline, polynomial splines with or without knots, quadratic polynomial,  and ARIMA.
}
\examples{
KK_28 <- readxl::read_excel("F:/Dyn4cast/R/data/Data.xlsx") # Nigeria COVID-19 data
KK_28$Date <- as.Date(KK_28$Date, format = '\%m/\%d/\%Y') # The date is reformatted
Dss <- seq(KK_28$Date[1], by = "day", length.out = length(KK_28$Case)) #data length for forecast
lastdayfo21 <- Dss[length(Dss)] # The maximum length
Data <- KK_28[KK_28$Date <= lastdayfo21 - 28, ] # desired length of forecast
BREAKS <- c(70, 131, 173, 228, 274) # The default breaks for the data
DynamicForecast(Data = Data, BREAKS = BREAKS, MaximumDate = "2021-02-10")

KK_14 <- readxl::read_excel("F:/Dyn4cast/R/data/Data.xlsx")
KK_14$Date <- as.Date(KK_14$Date, format = '\%m/\%d/\%Y')
Dss <- seq(KK_14$Date[1], by = "day", length.out = length(KK_14$Case))
lastdayfo21 <- Dss[length(Dss)]
Data <- KK_14[KK_14$Date <= lastdayfo21 - 14, ]
BREAKS = c(70, 131, 173, 228, 274)
DynamicForecast(Data = Data, BREAKS = BREAKS , MaximumDate = "2021-02-10")

}
