% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_plot.R
\name{estimate_plot}
\alias{estimate_plot}
\title{Plot of order of significance of estimated regression coefficients}
\usage{
estimate_plot(Model, limit)
}
\arguments{
\item{Model}{Estimated model for which the estimated coefficients would be plotted}

\item{limit}{Number of variables to be included in the coefficients plots}
}
\value{
The function returns a plot of the order of importance of the estimated coefficients
\item{\code{estimate_plot}}{The plot of the order of importance of estimated coefficients}
}
\description{
Plot of order of significance of estimated regression coefficients
}
