% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{gender}
\alias{gender}
\title{Create Gender Variable}
\usage{
gender(data)
}
\arguments{
\item{data}{data frame containing \strong{Age} and \strong{Sex} variables}
}
\value{
The \code{data.frame} with:
\item{\code{Gender}}{data frame with two additional variables.}
}
\description{
Often, there is need to differentiate between sex and gender. Many wonder if
there is any difference at all. This function will create clarity between
them.
}
\examples{
df <- data.frame(Age = c(49, 30, 44, 37, 29, 56),
 Sex = c("male", "female", "female", "male", "Prefer not to say",
  "Non-binary/third gender"))
 gender(df)
}
