% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_construction.R
\name{index_construction}
\alias{index_construction}
\title{Index Construction for estimation of \strong{Exposure} or \strong{Sensitivity}}
\usage{
index_construction(data)
}
\arguments{
\item{data}{Data frame of indicators of Exposure or Sensitivity. The data
frame must be numeric.}
}
\value{
A list with the following components:
\item{\code{Indexed data}}{\code{dataframe} of indices corresponding to the
supplied data.}
\item{\code{Index}}{A vector of indices representing the variable of
interest, either \strong{Exposure} or \strong{Sensitivity}.}
}
\description{
Vulnerability or to be vulnerable means the state or quality of being
susceptible to physical or emotional harm, damage, or attack, usually
indicating a lack of defense or protection, making someone or some systems
more likely to be affected by external factors or threats. Therefore,
vulnerability index is a quantitative or standardized framework of  such
state or quality which then makes comparisons between households,
communities or systems possible. The index is made up of three main
components: exposure, sensitivity and adaptive capacity. Each component
has multiple indicators from wide ranges including social, medical,
psychological and various extreme events like floods, drought, earthquakes
etc. This function is for conversion of indicators exposure and
sensitivity into a vector of index through normalization and weighting.
The resulting index from each of the component is then combined via an
appropriate model into vulnerability index.
}
\examples{
library(readr)
garrett_data <- data.frame(garrett_data)
index_construction(garrett_data)
}
