% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_scaled_logit.R
\name{inv_scaled_logit}
\alias{inv_scaled_logit}
\title{Exponential Values after One-Sided Response Integer Variable Forecasting}
\usage{
inv_scaled_logit(x, lower, upper)
}
\arguments{
\item{x}{The forecast values from constrained forecast package. Please specify the appropriate column containing the forecast values.}

\item{lower}{Lower limits of the forecast values}

\item{upper}{Upper limits of the forecast values}
}
\description{
This package is used to estimate exponential lower (80\% and 95\%) and upper (80\% and 95\%) values from the outcome of the \code{scaled_logit} package. The exponentiation ensures that the forecast does not go beyond the upper and lower limits of the base data.
}
\examples{
x <- 1:35
lower <- 1
upper <- 35
inv_scaled_logit(x = x, lower = lower, upper = upper)
}
