% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invscaledlogit.R
\name{invscaledlogit}
\alias{invscaledlogit}
\title{Exponential Values after One-Sided Response Integer Variable Forecasting}
\usage{
invscaledlogit(x, x3, lower, upper)
}
\arguments{
\item{x}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.}

\item{x3}{The forecast values from constrained forecast package. Please
specify the appropriate column containing the forecast values.}

\item{lower}{Lower limits of the forecast values}

\item{upper}{Upper limits of the forecast values}
}
\description{
This function is used to estimate exponential lower (80\% and 95\%) and upper
(80\% and 95\%) values from the outcome of the \code{scaledlogit} function.
The exponentiation ensures that the forecast does not go beyond the upper
and lower limits of the base data.
}
\examples{
x3 <- 1:35
lower <- 1
upper <- 35
invscaledlogit(x3 = x3, lower = lower, upper = upper)
}
