% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linearsystems.R
\name{Linearsystems}
\alias{Linearsystems}
\alias{linearsystems}
\title{Linear Model and various Transformations for Efficiency}
\arguments{
\item{y}{Vector of the dependent variable. This must be numeric.}

\item{x}{Data frame of the explanatory variables.}
}
\value{
A list with the following components:
\item{\code{Linear}}{The full estimates of the Linear Model.}
\item{\code{Semilog}}{The full estimates of the Semilog Model. Here the independent variable(s) is/are log-transformed.}
\item{\code{Growth}}{The full estimates of the Growth Model. Here the dependent variable(s) is/are log-transformed.}
\item{\code{Double Log}}{The full estimates of the Semilog Model. Here the both the dependent and independent variables are log-transformed.}
\item{\code{Quardratic}}{The full estimates of the Quadratic Model. Here the square of \emph{numeric} independent variable(s) is/are included as independent variables.}
\item{\code{Inverse y}}{The full estimates of the Inverse Model. Here the dependent variable is inverse-transformed (1/y).}
\item{\code{Square root y}}{The full estimates of the Square root Model. Here the dependent variable is square root-transformed (y^1/2).}
\item{\code{Model Table}}{Formatted Table of the coefficient estimates of all the models}
\item{\code{Prediction plots}}{Plots o the prediction from each of the model.}
\item{\code{Summary of numeric variables}}{of the dataset.}
\item{\code{Summary of character variables}}{of the dataset.}
}
\description{
The linear model still remains a reference point towards advanced modeling of some datasets as foundation for \strong{Machine Learning}, \strong{Data Science} and \strong{Artificial Intelligence} in spite of some of her weaknesses. The major task in \strong{modeling} is to compare various models before a selection is made for one or for advanced modelling. Often, some trial and error methods are used to decide which model to select. This is where this function is unique. It helps to estimate seven different linear models and provide their coefficients in a Table for quick comparison so that time and energy are saved. The interesting thing about this function is the simplicity, and it is a \emph{one line} code.
}
\examples{
y = linearsystems$MKTcost
x <- select(linearsystems, -MKTcost)
Linearsystems(y = y, x = x)
}
