% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mallows_Cp.R
\name{Mallows_Cp}
\alias{Mallows_Cp}
\title{Computation of Mallows Cp}
\usage{
Mallows_Cp(
  formula,
  data = list(),
  model = TRUE,
  x = FALSE,
  y = FALSE,
  var.full = 0,
  contrasts = NULL,
  verbose = FALSE
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. The details of model specification are given below.}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which Mallows Cp is called from.}

\item{model}{logical. If TRUE the corresponding components of the fit (the model frame, the model matrix, the response).}

\item{x}{logical. If TRUE the corresponding components of the fit (the model frame, the model matrix, the response).}

\item{y}{logical. If TRUE the corresponding components of the fit (the model frame, the model matrix, the response).}

\item{var.full}{the value of variance to be used in the denominator of the Mallows Cp, if 0 the variance estimated from the full model is used.}

\item{contrasts}{an optional list. See the contrasts.arg of model.matrix.default.}

\item{verbose}{if TRUE warnings are printed.}
}
\value{
\item{\code{Mallows Cp}}{The computed value of the model's Mallows Cp.}
}
\description{
Computation of Mallows Cp
}
