% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdi.R
\name{plot_mdi}
\alias{plot_mdi}
\title{Plots of Multidimensional Index Measures}
\usage{
plot_mdi(data, kala, dma, factor = NULL)
}
\arguments{
\item{data}{\verb{Data frame} of Multidimensional Index measures which is an
object from \code{mdi}}

\item{kala}{color palette with at least 15 colors but must be equal or higher
than the number of options in the factor argument}

\item{dma}{number of \code{Dimensions} involved in the computation of
Multidimensional Index measures.}

\item{factor}{the optional grouping factor used in the computation measures.
If not supplied only the national plots will be produced irrespective of
whether the factor was used in the computation.}
}
\value{
A list of the following plots:
\item{\code{Multidimensional index}}{plot.}
\item{\code{Deprivation Score}}{plot.}
\item{\code{Adjusted incidence}}{plot.}
\item{\code{Intensity}}{plot.}
\item{\code{Average deprivation among the deprived}}{plot.}
\item{\code{Contribution of each Dimension}}{plot.}
\item{\code{combined dimensions}}{plot.}
\item{\code{national}}{plot.}
\item{\code{combined dimensions of national}}{plot.}
}
\description{
Plots of Multidimensional Index Measures
}
\examples{
# Not run, uncomment to run
# library(MPI)
# data("examplePovertydf")
# data <- examplePovertydf
# dm <- list(d1 = c("Child.Mortality", "Access.to.health.care"),
#            d2 = c("Years.of.education", "School.attendance", "School.lag"),
#            d3 = c("Cooking.Fuel", "Access.to.clean.source.of.water",
#                   "Access.to.an.improve.sanatation", "Electricity",
#                   "Housing.Materials", "Asset.ownership"))
# dp <- mdi(data, dm, Factor = "Region")
# library(MetBrewer)
# kala <- met.brewer("OKeeffe1", 15, type = "continuous")
# dma <- 3
# plot_mdi(dp$MDI, kala, dma, "Region")
}
