% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdpi.R
\name{plot_mdpi}
\alias{plot_mdpi}
\title{Plots of Multidimensional Poverty Measures}
\usage{
plot_mdpi(data, kala, factor = NULL)
}
\arguments{
\item{data}{\verb{Data frame} of Multidimensional Poverty measures from \code{mdpi}}

\item{kala}{color palette with at least 15 colors but must be equal or higher
than the number of options in the factor argument}

\item{factor}{the optional grouping factor used in the computation measures.
If not supplied only the national plots will be produced irrespective of
whether the factor was used in the computation.}
}
\value{
A list of the following plots:
\item{\code{Multidimensional poverty index}}{plot.}
\item{\code{Deprivation Score}}{plot.}
\item{\code{Adjusted incidence of poverty}}{plot.}
\item{\code{Intensity of poverty}}{plot.}
\item{\code{Average deprivation among the deprived}}{plot.}
\item{\code{Contribution of each Dimension}}{plot.}
\item{\code{combined dimensions}}{plot.}
\item{\code{national}}{plot.}
\item{\code{combined dimensions of national}}{plot.}
}
\description{
Plots of Multidimensional Poverty Measures
}
\examples{
# Not run, uncomment to run
# library(MPI)
# data("examplePovertydf")
# data <- examplePovertydf
# dm <- list(d1 = c("Child.Mortality", "Access.to.health.care"),
#            d2 = c("Years.of.education", "School.attendance", "School.lag"),
#            d3 = c("Cooking.Fuel", "Access.to.clean.source.of.water",
#                   "Access.to.an.improve.sanatation", "Electricity",
#                   "Housing.Materials", "Asset.ownership"))
# dp <- mdpi(data, dm, Factor = "Region")
# library(MetBrewer)
# kala <- met.brewer("OKeeffe1", 15, type = "continuous")
# plot_mdpi(dp$MDPI, kala, "Region")
}
