% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_square.R
\name{predict.square}
\alias{predict.square}
\title{Predictive Residual Sum of Squares}
\arguments{
\item{object}{Model Estimated}

\item{verbose}{Outputs details argument}
}
\description{
Predictive Residual Sum of Squares
}
