% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksummary.R
\name{quicksummary}
\alias{quicksummary}
\alias{Quicksummary}
\title{Quick Formatted Summary of Machine Learning Data}
\usage{
quicksummary(
  x,
  Type,
  Cut = deprecated(),
  Up = deprecated(),
  Down = deprecated(),
  ci = 0.95
)
}
\arguments{
\item{x}{The data to be summarised. Only numeric data is allowed.}

\item{Type}{The type of data to be summarised. There are two options here 1
or 2, 1 = \code{Continuous} and 2 = \code{Likert-type}}

\item{Cut}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{Up}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{Down}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{ci}{Confidence interval which is defaults to 0.95.}
}
\value{
The function returns formatted tables of the Quick summary
\item{\code{Summary}}{List of two \code{data.frames}}
}
\description{
There is increasing need to make user-friendly and production ready Tables
for machine learning data. This function is simplified and quick summary;
and the output is a formatted table. This is very handy for those who
do not have the time to write codes for user-friendly summaries.
}
\examples{
library(tidyverse)
# Likert-type data
quicksummary(x = Quicksummary, Type = 2)

# Continuous data
x <- select(linearsystems, 1:6)
quicksummary(x = x, Type = 1)
}
