% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_likert.R
\name{relative_likert}
\alias{relative_likert}
\title{Convert Likert Data to Relative Scores and knowledge-based \strong{Adaptive Capacity}}
\usage{
relative_likert(data, Likert = NULL, Ranks = NULL, Option = "text")
}
\arguments{
\item{data}{Data frame of likert data either in \strong{text} or \strong{scores}.}

\item{Likert}{Vector of likert-type factors in \strong{descending order} as in the
data frame which must be given if the data frame is in \strong{text}.}

\item{Ranks}{Optional vector of number of levels which is required if the
data frame is in scores rather than text. There are only four choices
i.e. 3, 5, 7, 9.}

\item{Option}{Optional vector indicating whether the data frame is in text
or scores format. Defaults to \strong{text} if not given.}
}
\value{
A list with the following components:
\item{\code{lik_num}}{\code{dataframe} of likert scores}
\item{\code{lik_rate}}{\code{dataframe} of relative likert scores}
\item{\code{lik_sum}}{\code{dataframe} summary of relative scores based on factors.}
\item{\code{lik_col}}{A vector of indices for \strong{Adaptive Capacity}.}
}
\description{
Adaptive capacity refers to the ability of systems—biological, social, or
institutional—to adjust to environmental changes, capitalize on emerging
opportunities, and mitigate potential threats in order to preserve essential
functions. In the context of climate change, adaptive capacity denotes the
competence of social-ecological systems to cope with present variability
and prepare for uncertain future conditions.

From a machine learning perspective, adaptive capacity is closely linked to
the system’s ability to process large-scale, heterogeneous data sources,
identify patterns, and support the development of predictive models and
adaptive strategies. Key components of adaptive capacity include access to
relevant and reliable information, computational infrastructure, financial
and human capital, and strong social and institutional networks.

Machine learning can enhance adaptive capacity by enabling dynamic learning
from historical and real-time data, improving climate risk assessments, and
optimizing adaptation strategies. Moreover, the iterative nature of model
refinement and feedback integration mirrors the learning processes
inherent in adaptive systems. Thus, adaptive capacity in this context
involves not only the ability to design and implement effective
interventions but also to learn from outcomes and continuously update
strategies in light of new data and evolving conditions.

This function is for knowledge-based \strong{Adaptive Capacity}. The indices from
the various knowledge areas like Awareness, availability, affordability,
accessibility, benefits, adequacy, usage, effectiveness, etc can be
obtained individually and converted to adaptive capacity. Adaptive
capacity based on financial and human capital, strong social and
institutional networks can be obtained from \code{model_factors()}.
}
\examples{
library(readr)
garrett_data <- data.frame(garrett_data)
relative_likert(garrett_data, Ranks = 3, Option = "sccore")
relative_likert(garrett_data, Ranks = 5, Option = "sccore")
relative_likert(garrett_data, Ranks = 7, Option = "sccore")
relative_likert(garrett_data, Ranks = 9, Option = "sccore")

relative_likert(Quicksummary, Ranks = 5, Option = "sccore")

library(tidyverse)
data_l <- garrett_data \%>\%
pivot_longer(cols = everything()) \%>\%
 mutate(value = case_when(value == 5 ~ "Serious constraint",
                          value == 4 ~ "Constraint",
                          value == 3 ~ "Not certain it is a constraint",
                          value == 2 ~ "Not a constraint",
                          value == 1 ~ "Not a serious constraint",
                          .default = "None")) \%>\%
 group_by(name) \%>\%
 mutate(row = row_number()) \%>\%
 pivot_wider(names_from = name, values_from = value) \%>\%
 select(-row) \%>\%
 unnest(cols = everything())

 ranking <- c("Serious constraint", "Constraint",
"Not certain it is a constraint", "Not a constraint",
"Not a serious constraint")

 relative_likert(data_l, Likert = ranking)
}
