% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_logit.R
\name{scaled_logit}
\alias{scaled_logit}
\title{Scale Parameter for Integer Modelling and Forecast}
\usage{
scaled_logit(x, lower, upper)
}
\arguments{
\item{x}{The parameter to be scaled, which should normally be the fitted values from supported models. The scaled parameter is used mainly for constrained forecasting of a response variable >> positive {0 - inf} or << negative {-inf - 0}). The scaling also involves log transformation of the parameter}

\item{lower}{Integer or variable representing the lower limits for the scaling (-inf or 0)}

\item{upper}{Integer or variable representing the upper limits for the scaling (0 or inf)}
}
\value{

}
\description{
Scale Parameter for Integer Modelling and Forecast
}
\examples{
X <- 0:35
lower <- 0
upper <- 35
scaled_logit(x = x, lower = lower, upper = upper)

}
