% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_Dyn.R
\name{Dyn4cast-deprecated}
\alias{Dyn4cast-deprecated}
\alias{Model_factors}
\title{Deprecated functions in Dyn4cast}
\usage{
Model_factors(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
None, function now defunct.
}
\description{
These functions have been removed (defunct) as indicated against their names.
}
\examples{
# Non
}
