% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dyn4cast-package.R
\docType{package}
\name{Dyn4cast-package}
\alias{Dyn4cast}
\alias{Dyn4cast-package}
\title{Dyn4cast: Dynamic Forecast of Five Models and their Ensembles}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimates, predict and forecast five models and their Ensembles. The recognized models are lm, smooth spline, polynomial splines with or without knots, quadratic polynomial, and ARIMA. The package is so far stable and takes `daily, monthly and yearly datasets for now`. The robust output include the models' estimates, time-varying forecasts and plots based on themes from ggplot. The main attraction of this package is the use of the newly introduced _equal sample size (time, trend) forecast_.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JobNmadu/Dyn4cast}
  \item Report bugs at \url{https://github.com/JobNmadu/Dyn4cast/issues}
}

}
\author{
\strong{Maintainer}: Job Nmadu \email{job.nmadu@gmail.com} (\href{https://orcid.org/0000-0002-1320-8957}{ORCID})

}
\keyword{internal}
