% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dyn4cast-package.R
\docType{package}
\name{Dyn4cast-package}
\alias{Dyn4cast}
\alias{Dyn4cast-package}
\title{Dyn4cast: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
