% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dyn4cast-package.R
\docType{package}
\name{Dyn4cast-package}
\alias{Dyn4cast}
\alias{Dyn4cast-package}
\title{Dyn4cast: Dynamic Modeling and Machine Learning Environment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimates, predict and forecast dynamic models as well as Machine Learning metrics which assists in model selection for further analysis. The package also have capabilities to provide tools and metrics that are useful in machine learning and modeling. For example, there is quick summary, percent sign, Mallow's Cp tools and others. The ecosystem of this package is analysis of economic data for national development. The package is so far stable and has high reliability and efficiency as well as time-saving.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JobNmadu/Dyn4cast}
  \item Report bugs at \url{https://github.com/JobNmadu/Dyn4cast/issues}
}

}
\author{
\strong{Maintainer}: Job Nmadu \email{job.nmadu@gmail.com} (\href{https://orcid.org/0000-0002-1320-8957}{ORCID})

}
\keyword{internal}
