% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicForecast.R
\name{DynamicForecast}
\alias{DynamicForecast}
\title{DynamicForecast}
\usage{
DynamicForecast(Data, BREAKS, Date)
}
\arguments{
\item{Data}{A data.frame}

\item{BREAKS}{A vector}

\item{Date}{A date}
}
\value{
results
}
\description{
DynamicForecast
}
\examples{
Data <- readxl::read_excel("F:/Dyn4cast/R/data/Data.xlsx")
Data$Date <- as.Date(Data$Date, format = '\%m/\%d/\%Y')
DynamicForecast(Data = Data, BREAKS = c(70, 131, 173, 228, 274),
 Date = "2021-02-10")
}
