% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_plot.R
\name{estimate_plot}
\alias{estimate_plot}
\title{Plot of Order of Significance of Estimated Regression Coefficients}
\usage{
estimate_plot(model, limit)
}
\arguments{
\item{model}{Estimated model for which the estimated coefficients would be
plotted}

\item{limit}{Number of variables to be included in the coefficients plots}
}
\value{
The function returns a plot of the order of importance of the
estimated coefficients
\item{\code{estimate_plot}}{The plot of the order of importance of estimated
coefficients}
}
\description{
This function provides graphic displays of the order of significance
estimated coefficients of models. This would assists in accessing models so
as to decide which can be used for further analysis, prediction and policy
consideration.
}
