% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_scaled_logit.R
\name{inv_scaled_logit}
\alias{inv_scaled_logit}
\title{Exponential Values  after Integer Forecasting}
\usage{
inv_scaled_logit(x, lower, upper)
}
\arguments{
\item{x}{The forecast values from constrained forecast package. Please specify the appropriate column (x\link{ , 1}) containing the forecast values.}

\item{lower}{Lower limits of the forecast values}

\item{upper}{Upper limits of the forecast values}
}
\value{

}
\description{
Exponential Values  after Integer Forecasting
}
\examples{
X <- 0:35
lower <- 0
upper <- 35
inv_scaled_logit(x = X, lower = lower, upper = upper)
}
