% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MallowsCp.R
\name{MallowsCp}
\alias{MallowsCp}
\title{Computation of MallowsCp}
\usage{
MallowsCp(Model, y, x, type, Nlevels = 0)
}
\arguments{
\item{Model}{The estimated \strong{model} from which the Mallows Cp would be computed}

\item{y}{The vector of the \strong{LHS} variable of the estimated model}

\item{x}{The matrix of the \strong{RHS} variable of the estimated model. Note that if the model adds additional factor variables into the output, then the number of additional factors \code{Nlevels} is required otherwise the computed Cp would be biased.}

\item{type}{The type of model for which Cp would be computed broadly divided in to linear and non-linear. If type is non-linear, specify the name of the model. Supported models are \code{LM}, \code{ALM}, \code{GLM} for binary based models, \code{N-LM} (not linear for models not clearly defined as linear or non-linear especially some of the essemble models that are merely \strong{computed} not \strong{estimated}) or \code{nls} for other non linear models, \code{ARDL}, \code{SMOOTH} for \strong{smooth.spline}, \code{SPLINE} for bs spline models, \code{ARIMA} and \code{plm}.}

\item{Nlevels}{Optional number of additional variables created if the model has categorical variables that generates additional dummy variables during estimation or the number of additional variables created if the model involves interaction terms.}
}
\value{
A list with the following components
\item{\code{MallowsCp}}{of the Model.}
}
\description{
This function is one of the numerous metrics used to assess and compare linear based models. The measure gives an approximate number of explanatory variables that should be in the model. The function inherits \code{residuals} from the estimated model.
}
\examples{
library(Dyn4cast)
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
x <- gl(2, 10, 20, labels = c("Ctl","Trt"))
y <- c(ctl, trt)
Model <- lm(y ~ x)
Type <- "LM"
MallowsCp(Model = Model, y = y, x = x, type = Type, Nlevels = 0)
}
