% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MallowsCp.R
\docType{package}
\name{MallowsCp}
\alias{MallowsCp}
\title{Computation of Mallows_Cp}
\usage{
MallowsCp(Model, y, x, Nlevels = NULL)
}
\arguments{
\item{Model}{The estimated \strong{model} from which the Mallow's Cp would be computed}

\item{y}{The vector of the \strong{LHS} variable of the estimated model}

\item{x}{The matrix of the \strong{RHS} variable of the estimated model. Note that if the model adds additional factor variables into the output, then the number of additional factors \code{Nlevels} is required otherwise the computed Cp would be biased.}

\item{Nlevels}{Optional number of additional variables created if the model has categorical variables that generates additional dummy variables during estimation}
}
\value{
A list with the following components:
\item{\code{Mallow's Cp}}{of the model.}
}
\description{
Compute the Mallow's Cp of the model
}
\details{
This function is one of the numerous metrics used to assess and compare linear-based models. The measure gives an approximate number of explanatory variables that should be in the model.
}
\examples{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
x <- gl(2, 10, 20, labels = c("Ctl","Trt"))
y <- c(ctl, trt)
Model <- lm(y ~ x)
MallowsCp(Model = Model, y = y, x = x, Nlevels = NULL)

}
