% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpi.R
\name{mdpi}
\alias{mdpi}
\title{Sequential Computation of Dynamic Multidimensional Poverty Indices (MDPI)}
\usage{
mdpi(
  data,
  dm,
  Bar = 0.4,
  id_addn = NULL,
  Factor = NULL,
  plots = NULL,
  id = c("Health", "Education", "Living standard"),
  id_add = "Social security",
  id_add1 = "Employment and Income"
)
}
\arguments{
\item{data}{\verb{data frame} containing all the variables for the computation.
Note that the variables to be used for the computation must be coded \verb{(0,1)}.}

\item{dm}{list of vectors of \emph{indicators} making up each \emph{dimension} to be
computed}

\item{Bar}{a vector of cut-of used to divide the population into those in
the poverty category and those that are not. Defaults to 0.4 if not supplied.}

\item{id_addn}{a vector of additional dimensions to be used for the computation
up to a \emph{maximum of four}.}

\item{Factor}{a grouping factor for the computation which must be a variable
in the \emph{data}.}

\item{plots}{plots of the various measures. For this to be possible, the
number of options in the \code{Factor} argument must be less than 41. The default
is \code{NULL}. To produce, any character string will overwrite the default.}

\item{id}{a vector of the first three dimensions used in the computation
given as \strong{Health}, \strong{Education} and \strong{Living standard}.}

\item{id_add}{a vector of the fourth dimension in the computation given
as \strong{Social security}. Can be re-defined but never \code{NULL}.}

\item{id_add1}{a vector of the fifth dimension in the computation given
as \strong{Employment and Income}. Can be re-defined but never \code{NULL}.}
}
\value{
A list with the following components:
\item{\code{MDPI_p}}{Publication-ready table of the factor and national
MDPI prepared with \verb{summarymodels package}. Will not \emph{return} if only
national computation is carried out.}
\item{\code{MDPI}}{\verb{Data frame} of the factor and national MDPI. Will
not \emph{return} if only national computation is carried out.}
\item{\code{MDPI mean}}{\verb{Data frame} of the mean MDPI. Will not \emph{return} if
only national computation is carried out.}
\item{\code{MDPI SD}}{\verb{Data frame} of the SD of MDPI. Will not \emph{return} if
only national computation is carried out.}
\item{\code{national}}{\verb{Data frame} of national MDPI with mean and SD.}
\item{\code{dimensions}}{\verb{Data frame} of the scores for each dimension in
the computation.}
\item{\code{Score}}{\verb{Data frame} of the scores for each indicator in the
computation.}
}
\description{
This function computes the indices and all associated measures of
multidimensional poverty sequentially in a dynamic way. Sequentially
the function computes \emph{Incidence of poverty (H = q / n)},
\emph{Adjusted incidence of poverty (H / (q / D))}, \emph{Deprivation Score} of each
dimension in the computation, \emph{Intensity of poverty (A)},
\emph{Multidimensional poverty index (MDPI = H * A)}, the \emph{Contribution} in
\% of each of the dimensions to MDPI, and
\emph{Average deprivation among the deprived (A * D)}. Dynamically, it
computes the various indices for between three and nine \code{dimensions (D)}.
The first five dimensions included in the computations are \emph{Health},
\emph{Education}, \emph{Living standard}, \emph{Social security} and,
\emph{Employment and Income} depending on the choice of the user. Four
additional dimensions can be included in the computations. The
computations are carried out either for the \verb{national sample data} or
can be dis-aggregated based on \verb{grouping factors}, like region, sex,
gender, marital status or any suitable one. The cut-off mark
demarcating \code{poor (q)} and \code{non-poor (n-q)} members in the \code{sample (n)}
is defaulted to \code{0.4} but can be varied as may be dictated by the
interests or the need for the computation. The computations are in
line with various procedures already outlined in literature starting
with the work of Alkire et. al, (2015) but has been expanded from
three dimensions to nine. Each dimension is given \verb{equal weight} in
the computation but all indicators are weighted in line with
existing guidelines in Alkire & Foster (2011) and Alkire & Santos
(2010). See also Alkire & Santos (2014) and Chan & Wong (2024).
}
\examples{
# Not run, uncomment to run
# library(MPI)
# data("examplePovertydf")
# data <- examplePovertydf
# dm <- list(d1 = c("Child.Mortality", "Access.to.health.care"),
#            d2 = c("Years.of.education", "School.attendance", "School.lag"),
#            d3 = c("Cooking.Fuel", "Access.to.clean.source.of.water",
#                   "Access.to.an.improve.sanatation", "Electricity",
#                   "Housing.Materials", "Asset.ownership"))
# mdpi(data, dm, plots = "t", Factor = "Region")
#
# library(mpitbR)
# data <- subset(syn_cdta)
# data <- na.omit(data)
# dm <- list(d1 = c("d_nutr","d_cm"),
#            d2 = c("d_satt","d_educ"),
#            d3 = c("d_elct","d_sani","d_wtr","d_hsg","d_ckfl","d_asst"))
# mdpi(data, dm, plots = "t", Factor = "region")

}
\references{
Alkire, S. & Foster, J. (2011). Counting and Multidimensional Poverty
Measurement. Journal of Public Economics 95(7-8): 476–87.
https://doi.org/10.1016/j.jpubeco.2010.11.006.

Alkire, S., Foster, J. E., Seth, S., Santos, M. E., Roche, J., & Ballon, P.
(2015). Multidimensional poverty measurement and analysis. Oxford University
Press.

Alkire, S. & Santos, M. E. (2010).  Acute Multidimensional Poverty: A New
Index for Developing Countries. Oxford Poverty and Human Development
Initiative (OPHI) Working Paper No. 38.

Alkire, S. & Santos, M. E. (2014). Measuring Acute Poverty in the Developing
World: Robustness and Scope of the Multidimensional Poverty Index. World
Development 59:251-274. https://doi.org/10.1016/j.worlddev.2014.01.026.

Siu Ming Chan & Hung Wong (2024): Measurement and determinants of
multidimensional poverty: the case of Hong Kong, Journal of Asian Public
Policy, DOI: 10.1080/17516234.2024.2325857
}
