% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_factors.R
\name{Model_factors}
\alias{Model_factors}
\title{Latent Variables Recovery from Factor Loadings}
\arguments{
\item{data}{An \verb{R object} obtained from exploratory factor analysis (EFA) using the \code{fa} function in \code{psych} package.}

\item{DATA}{A \code{data.frame}, the raw data used to carry out the parallel analysis to obtain \code{data} object.}
}
\value{
A list with the following components:
\item{\code{Load_numeric}}{Table of latent factors and the variables loadings.}
\item{\code{Load_char}}{Table of latent factors and their variables.}
}
\description{
This function retrieves the latent factors and their variable loadings which can be used as \code{R} objects to perform other analysis.
}
\examples{
library(psych)
Data <- Quicksummary
GGn <- names(Data)
GG <- ncol(Data)
GGx <- c(paste0('x0', 1:9), paste("x", 10:ncol(Data), sep = ""))
names(Data) <- GGx
lll <- fa.parallel(Data, fm = 'minres', fa = 'fa')
dat <- fa(Data, nfactors = lll[["nfact"]], rotate = "varimax",fm="minres")

DD <- Model_factors(data = dat, DATA = Data)

factorM <- DD$Load_numeric
factorC <- DD$Load_char
}
