% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factors.R
\name{model_factors}
\alias{model_factors}
\title{Latent Factors Recovery from Variables Loadings}
\usage{
model_factors(data, DATA, RC = "No")
}
\arguments{
\item{data}{An \verb{R object} obtained from exploratory factor analysis (EFA)
using the \code{fa} function in \code{psych} package.}

\item{DATA}{A \code{data.frame}, the raw data used to carry out the parallel
analysis to obtain \code{data} object.}

\item{RC}{Optional factor indicating whether resilience capacity is to be
estimated but defaults to \code{NULL} once the number of variables in the data
is not sufficient, i.e. < 20. To estimate, turn it to "Yes".}
}
\value{
A list with the following components:
\item{\code{Loadings data}}{\code{dataframe} of the factor loadings from the
data.}
\item{\code{Factors extracted}}{\code{dataframe} of retrieved latent factors.}
\item{\code{factored data}}{\code{dataframe} of latent data based the product of
recovered latent factors and the on raw data.}
\item{\code{Factors list}}{A list of vectors of individual latent factors
recovered from the data. However, to make it usable, the vector should
be \code{bind} with the names of the variables in the data and the
\code{NA} removed.}
\item{\code{Resilence capacity}}{A vector of the resilience capacity if the
data is prepared for that otherwise NULL.}
}
\description{
This function retrieves the latent factors and their variable loadings which
can be used as \code{R} objects to perform other analysis.
}
\examples{
library(psych)
library(readr)
Data <- Quicksummary
GGn <- names(Data)
GG <- ncol(Data)
GGx <- c(paste0('x0', 1 : 9), paste("x", 10 : ncol(Data), sep = ""))
names(Data) <- GGx
lll <- fa.parallel(Data, fm = "minres", fa = "fa")
dat <- fa(Data, nfactors = lll[["nfact"]], rotate = "varimax",fm = "minres")

model_factors(data = dat, DATA = Data)

}
