% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Percent.R
\docType{package}
\name{Percent}
\alias{Percent}
\title{Attach Per Cent Sign to Data}
\usage{
Percent(Data, Type, digits = 2, format = "f", ...)
}
\arguments{
\item{Data}{The Data which the percent sign is to be affixed}

\item{Type}{The type of data. The default arguments are \emph{Value} for single numeric data of \emph{Frame} for a numeric vector data}

\item{digits}{Number of decimal points for the output}

\item{format}{The format of the output which is internal and the default is a character factor}

\item{...}{Additional arguments that may be passed to the function}
}
\value{

}
\description{
Affix the graphic percentage sign (\%) on a vector of numbers
}
\details{
This function is a wrapper for easy affixing of the per cent sign (\%) to a value or a vector of values.
}
\examples{
Data <- c(1.2, 0.5, 0.103, 7, 0.1501)
Percent(Data = Data, Type = "Frame")  # Value, Frame

Data <- 1.2
Percent(Data = Data, Type = "Value")  # Value, Frame

}
