% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdi.R
\name{plot_mdi}
\alias{plot_mdi}
\title{Plots of Multidimensional Index Measures}
\usage{
plot_mdi(data, kala, dma, factor = NULL)
}
\arguments{
\item{data}{\verb{Data frame} of Multidimensional Index measures which is an
object from \code{mdi}}

\item{kala}{color palette with at least 15 colors but must be equal or higher
than the number of options in the factor argument}

\item{dma}{number of \code{Dimensions} involved in the computation of
Multidimensional Index measures.}

\item{factor}{the optional grouping factor used in the computation measures.
If not supplied only the national plots will be produced irrespective of
whether the factor was used in the computation.}
}
\value{
A list of the following plots:
\item{\code{Multidimensional index}}{plot.}
\item{\code{Deprivation Score}}{plot.}
\item{\code{Adjusted incidence}}{plot.}
\item{\code{Intensity}}{plot.}
\item{\code{Average deprivation among the deprived}}{plot.}
\item{\code{Contribution of each Dimension}}{plot.}
\item{\code{combined dimensions}}{plot.}
\item{\code{national}}{plot.}
\item{\code{combined dimensions of national}}{plot.}
}
\description{
Plots of Multidimensional Index Measures
}
\examples{
# data from `mpitbR` package
data <- mdpi2
dm <- list(d1 = c("d_nutr","d_cm"),
           d2 = c("d_satt","d_educ"),
           d3 = c("d_elct","d_sani","d_wtr","d_hsg","d_ckfl","d_asst"))
dp <-  mdi(data, dm, plots = "t")
library(MetBrewer)
kala <- met.brewer("OKeeffe1", 20, type = "continuous")
dma <- 3
plot_mdi(dp$national, kala, dma)
}
